added: category update and delete functionality

+++ src/services/category.service.ts
@@ -1,4 +1,5 @@
 import { env } from "@/env";
+import { cookies } from "next/headers";
 
 export const getAllCategory = async (page: number, limit: number) => {
   try {
@@ -37,11 +38,48 @@ export const getAllCategory = async (page: number, limit: number) => {
   }
 };
 
+export const updateCategoryStatus = async (id: string, status: boolean) => {
+  try {
+    const cookieStore = await cookies();
+    const sessionToken = cookieStore.get("better-auth.session_token")?.value;
+
+    if (!sessionToken) return null;
+
+    const res = await fetch(`${env.BACKEND_URL}/api/category/${id}`, {
+      method: "PATCH",
+      headers: {
+        "Content-Type": "application/json",
+        Cookie: `better-auth.session_token=${sessionToken}`,
+      },
+      body: JSON.stringify({ isActive: status }),
+      cache: "no-store",
+    });
+
+    if (!res.ok) {
+      throw new Error(`Error: ${res.statusText}`);
+    }
+
+    const updateStatus = await res.json();
+
+    return updateStatus.data;
+  } catch (err) {
+    console.log(err);
+    throw err;
+  }
+};
+
 export const deleteCategory = async (id: string) => {
   try {
-    const res = await fetch(`${env.BACKEND_URL}api/category/${id}`, {
+    const cookieStore = await cookies();
+    const sessionToken = cookieStore.get("better-auth.session_token")?.value;
+
+    if (!sessionToken) return null;
+    const res = await fetch(`${env.BACKEND_URL}/api/category/${id}`, {
       method: "DELETE",
-      headers: { "Content-Type": "application/json" },
+      headers: {
+        "Content-Type": "application/json",
+        Cookie: `better-auth.session_token=${sessionToken}`,
+      },
     });
 
     if (!res.ok) {
